[Refactor] #29 주간 리스트 뷰에 단기예보 활용을 위한 리팩토링
주간 리스트 뷰에 단기예보 활용을 위한 리팩토링
 - WeeklyWeatherManager 에서 WeatherManager로 병합
